# openapi/agh.json
{
...
    "info": {
        "title": "AGH API",
        "version": "1.0.0"
    },
    "openapi": "3.0.3",
    "paths": {
        "/api/agents": {
            "get": {
                "operationId": "listAgents",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "agents": {
                                            "items": {
                                                "properties": {
                                                    "command": {
                                                        "type": "string"
                                                    },
                                                    "mcp_servers": {
                                                        "items": {
                                                            "properties": {
                                                                "args": {
                                                                    "items": {
                                                                        "type": "string"
                                                                    },
                                                                    "type": "array"
                                                                },
                                                                "command": {
                                                                    "type": "string"
                                                                },
                                                                "env": {
                                                                    "additionalProperties": {
                                                                        "type": "string"
                                                                    },
                                                                    "type": "object"
                                                                },
                                                                "name": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "required": ["command", "name"],
                                                            "type": "object"
                                                        },
                                                        "type": "array"
                                                    },
                                                    "model": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "permissions": {
                                                        "type": "string"
                                                    },
                                                    "prompt": {
                                                        "type": "string"
                                                    },
                                                    "provider": {
                                                        "type": "string"
                                                    },
                                                    "tools": {
                                                        "items": {
                                                            "type": "string"
                                                        },
                                                        "type": "array"
                                                    }
                                                },
                                                "required": ["name", "prompt", "provider"],
                                                "type": "object"
                                            },
                                            "type": "array"
                                        }
                                    },
                                    "required": ["agents"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Internal server error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "summary": "List all readable agent definitions",
                "tags": ["agents"],
                "x-agh-transports": ["http", "uds"]
            }
        },
        "/api/agents/{name}": {
            "get": {
                "operationId": "getAgent",
                "parameters": [
                    {
                        "description": "Agent name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "agent": {
                                            "properties": {
                                                "command": {
                                                    "type": "string"
                                                },
                                                "mcp_servers": {
                                                    "items": {
                                                        "properties": {
                                                            "args": {
                                                                "items": {
                                                                    "type": "string"
                                                                },
                                                                "type": "array"
                                                            },
                                                            "command": {
                                                                "type": "string"
                                                            },
                                                            "env": {
                                                                "additionalProperties": {
                                                                    "type": "string"
                                                                },
                                                                "type": "object"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "required": ["command", "name"],
                                                        "type": "object"
                                                    },
                                                    "type": "array"
                                                },
                                                "model": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "permissions": {
                                                    "type": "string"
                                                },
                                                "prompt": {
                                                    "type": "string"
                                                },
                                                "provider": {
                                                    "type": "string"
                                                },
                                                "tools": {
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "type": "array"
                                                }
                                            },
                                            "required": ["name", "prompt", "provider"],
                                            "type": "object"
                                        }
                                    },
                                    "required": ["agent"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Agent not found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Internal server error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "summary": "Get one agent definition by name",
                "tags": ["agents"],
                "x-agh-transports": ["http", "uds"]
            }
        },
...
    },
...
}